implementacion del nivel de confianza

+++ app/Database/Migrations/2024-08-23-160658_UsersTable.php
@@ -10,7 +10,7 @@ class UsersTable extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type' => 'SMALLINT',
+                'type' => 'INT',
                 'constraint' => 5,
                 'unsigned' => true,
                 'auto_increment' => true,
